feat: add contested compare engine use case
